MP4: 64-bit box sizes (extended size, size 0), co64 and all tracks' chunk offsets updated on save (error instead of overflow), mdhd version 1, sample rate and channels from the sample entry, duration of the longest sound track, 16-bit track/disk/tempo numbers, free form items of any owner kept, GIF/BMP covers

+++ MP4_Container.h
@@ -35,9 +35,9 @@ public:
 	virtual bool removeAtom(CAtlString atomID);
 	void remove();
 	bool removeAtom(CMP4Atom* atom);
-	void load(FILE *stream, u32 offset, u32 size);
+	void load(FILE *stream, u64 offset, u64 size);
 	virtual void save(FILE *stream);
-	u32 getSize();	
+	u64 getSize();	
 protected:
-	u32 _offset, _size;
+	u64 _offset, _size;
 };
